+++ bin/generate_build_number.dart
@@ -1,10 +1,8 @@
 import 'dart:io';
-import 'dart:math';
 
 import 'package:args/args.dart';
+import 'package:dart_test_tools/src/ci/build_number_generator.dart';
 import 'package:dart_test_tools/tools.dart';
-import 'package:pub_semver/pub_semver.dart';
-import 'package:pubspec_parse/pubspec_parse.dart';
 
 Future<void> main(List<String> args) => Github.runZoned(() async {
       final parser = ArgParser(
@@ -25,6 +23,12 @@ Future<void> main(List<String> args) => Github.runZoned(() async {
           help: 'The <width> of the patch version in the build number.',
         )
         ..addFlag(
+          'env',
+          abbr: 'E',
+          help: 'Set the BUILD_NUMBER environment variable '
+              'instead of the buildNumber output.',
+        )
+        ..addFlag(
           'help',
           abbr: 'h',
           negatable: false,
@@ -39,9 +43,11 @@ Future<void> main(List<String> args) => Github.runZoned(() async {
           return;
         }
 
-        await _main(
+        const generator = BuildNumberGenerator();
+        await generator(
           minorWidth: int.parse(options['minor'] as String),
           patchWidth: int.parse(options['patch'] as String),
+          asEnv: options['env'] as bool,
         );
       } on FormatException catch (e) {
         stderr
@@ -51,35 +57,3 @@ Future<void> main(List<String> args) => Github.runZoned(() async {
         exitCode = 1;
       }
     });
-
-Future<void> _main({
-  required int minorWidth,
-  required int patchWidth,
-}) async {
-  final pubspecFile = File('pubspec.yaml');
-  final pubspec = Pubspec.parse(
-    await pubspecFile.readAsString(),
-    sourceUrl: pubspecFile.uri,
-  );
-
-  Github.logInfo('Detected app version as ${pubspec.version}');
-  final Version(major: major, minor: minor, patch: patch) = pubspec.version!;
-
-  final buildNumber = _padFilled('major', major, 0) +
-      _padFilled('minor', minor, minorWidth) +
-      _padFilled('patch', patch, patchWidth);
-
-  Github.logInfo('Generated build number as $buildNumber');
-
-  await Github.env.setOutput('buildNumber', buildNumber);
-}
-
-String _padFilled(String name, int number, int width) {
-  if (width > 0 && number >= pow(10, width)) {
-    throw Exception(
-      'Version number does not fit! '
-      'Segment $name ($number) has more then $width digits',
-    );
-  }
-  return number.toRadixString(10).padLeft(width, '0');
-}