+++ lib/src/custom_lint/custom_lint_plugin.dart
@@ -1,5 +1,6 @@
 // This is the entrypoint of our custom linter
 import 'package:custom_lint_builder/custom_lint_builder.dart';
+import 'package:dart_test_tools/src/custom_lint/lints/freezed_classes_must_be_sealed.dart';
 import 'package:dart_test_tools/src/custom_lint/lints/src_library_not_exported.dart';
 
 import 'lints/no_self_package_imports.dart';
@@ -13,5 +14,6 @@ class _CustomLintPlugin extends PluginBase {
   List<LintRule> getLintRules(CustomLintConfigs configs) => [
         const NoSelfPackageImports(),
         const SrcLibraryNotExported(),
+        const FreezedClassesMustBeSealed(),
       ];
 }