+++ lib/src/custom_lint/lints/freezed_classes_must_be_sealed.dart
@@ -1,0 +1,46 @@
+import 'package:analyzer/dart/ast/ast.dart';
+import 'package:analyzer/error/listener.dart';
+import 'package:custom_lint_builder/custom_lint_builder.dart';
+import 'package:dart_test_tools/src/custom_lint/lints/fixes/add_sealed_keyword.dart';
+
+class FreezedClassesMustBeSealed extends DartLintRule {
+  static const _code = LintCode(
+    name: 'freezed_classes_must_be_sealed',
+    problemMessage: 'Freezed classes must be sealed.',
+    correctionMessage: 'Add the sealed keyword to the class.',
+  );
+
+  const FreezedClassesMustBeSealed() : super(code: _code);
+
+  @override
+  void run(
+    CustomLintResolver resolver,
+    ErrorReporter reporter,
+    CustomLintContext context,
+  ) {
+    context.registry.addClassDeclaration((node) {
+      if (node.isUnsealedFreezed) {
+        reporter.reportErrorForToken(_code, node.classKeyword);
+      }
+    });
+  }
+
+  @override
+  List<Fix> getFixes() => [
+        FreezedClassesMustBeSealedFix(),
+      ];
+}
+
+extension ClassDeclarationX on ClassDeclaration {
+  bool get isUnsealedFreezed {
+    final isFreezed = metadata.any(
+      (a) => a.name.name.toLowerCase() == 'freezed',
+    );
+    if (!isFreezed) {
+      return false;
+    }
+
+    final isSealed = sealedKeyword != null;
+    return !isSealed;
+  }
+}